refactor(worker/agent): extract runOne body to runOneSpawn
Move the body of `runOne` into a new file `worker/agent/opencode_spawn.go`
as `(r *Runner) runOneSpawn(...)`. `runOne` becomes a thin wrapper that
unconditionally delegates to `runOneSpawn`. Behaviour-preserving — every
existing call site, signature, and error path is identical to the pre-
change inline implementation.

Sets up Task 3.2-3 (dispatcher), which will replace the unconditional
delegation with a mode check (agent.Command == "opencode" AND
cfg.Opencode.Mode == "server" → runOneServer; otherwise → runOneSpawn).
Stage 2+ runOne* variants for other modes can be added without further
touching runner.go.

Helpers used by runOneSpawn (readOutput, expandExtraArgs,
substituteStringPlaceholders, detectBlockedArgs, filterClaudeCodeEnv,
tailStderr, blockedArgs, claudeCodeEnvWhitelist, stderrTailLen) stay in
runner.go because they are not opencode-specific — every future
runOne* variant will reuse them.
